backend/requirements-api.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
tflite-runtime==2.14.0
numpy==2.4.4
Pillow==12.1.1
psycopg2-binary==2.9.11
